Rename to textNodeUtils
