+++ src/pdf/text/textNodeUtils.ts
